requirements.txt
pillow==11.1.0
PyMuPDF==1.25.3
